tests/wpt/tests/fedcm/fedcm-abort-silent-mediation-not-permitted.https.html

<!DOCTYPE html>
<title>Federated Credential Management API network request tests.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {fedcm_test,
        fedcm_get_and_select_first_account,
        request_options_with_mediation_required} from './support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let controller = new AbortController();
  let test_options = request_options_with_mediation_required();
  test_options.signal = controller.signal;
  test_options.mediation = "silent";
  const cred = fedcm_get_and_select_first_account(t, test_options);
  controller.abort();
  return promise_rejects_dom(t, 'NetworkError', cred);
}, "Test the abort signal when request fails due to silent mediation being disallowed");
</script>
